Merged PR 8905: 43552 - Fix for data source and connector changes to json files not pushed to cosmos through pipeline
Compare-Object used to compare JSON file content VS cosmos records was not comparing all nested properties since Compare-Object has some limitations to compare nested objects.

As a fix, we are now converting json to string and then compare the strings to check if anything got changed. Note that there are some properties are configured to be ignored from cosmos side since they are not applicable for comparison.

Related work items: #43552