Merged PR 9074: Fix missing WebUrl introduced by changing the documents response shape
To support paging, we changed the shape of the documents response from the connector.

See MockConnectorData.cs for an example of the changed shape.

The code in IManageConnectorOperations.cs needed to change to match the new shape in order to find the place in the results to add WebUrl

Related work items: #44913