Merged PR 10140: Changes to include new OAuth Gettoken
Changes to include new OAuth Gettoken

I changed the object that we are using so that we are always using a new object that more closely represents the OAuth service response.
Also the OAuthService is not being used but I updated it to also have a new token type.  We might want to remove that because no on is using it at the moment.

Related work items: #50266