Merged PR 10186: updating user id and tenant id
Not sure what happened with the last change.  I updated connectors with a new user to hopefully help with test issues but when checked in the id and tenant were marked deleted and a new one was setup.

![image.png](https://tr-legal.visualstudio.com/0308fe1f-6035-44ba-8617-dddaffafe6f1/_apis/git/repositories/e31481b2-24e7-4514-b69b-dd19da98a03b/pullRequests/10186/attachments/image.png)

Concerning but we will keep our eyes on this

Related work items: #50476