Merged PR 13323: Add validation for IH token scopes for DMS flow.
Add validation for IH token scopes for DMS flow.

Validation added into ThoughtTraceOntologiesConnectorOperation (if scopes are not valid, no search is performed).
![image (5).png](https://tr-legal.visualstudio.com/0308fe1f-6035-44ba-8617-dddaffafe6f1/_apis/git/repositories/e31481b2-24e7-4514-b69b-dd19da98a03b/pullRequests/13323/attachments/image%20%285%29.png)

**Response before changes:**
The search was performed and a generic error message was returned due the failure when calling the IH permissions endpoint
![image (2).png](https://tr-legal.visualstudio.com/0308fe1f-6035-44ba-8617-dddaffafe6f1/_apis/git/repositories/e31481b2-24e7-4514-b69b-dd19da98a03b/pullRequests/13323/attachments/image%20%282%29.png)

**Response after changes:**
If scopes are not valid, no search is performed and a specific message exception including the list of required scopes is thrown.
![image (4).png](https://tr-legal.visualstudio.com/0308fe1f-6035-44ba-8617-dddaffafe6f1/_apis/git/repositories/e31481b2-24e7-4514-b69b-dd19da98a03b/pullRequests/13323/attachments/image%20%284%29.png)

Related work items: #298588